Wrote AlphaBeta, GameTree and Node_JumpSturdy classes
Three classes to help with the final alpha beta pruining algorithm. The Alpha Beta class will only do the Algorithm.
The Tree class shall manipulate and generate the Trees, which the Alpha Beta algortihm will analyse. The Node_JumpSturdy class will be used by the Tree class. I made a seperate Node class, so it is easier to work on it.