worked on zobrist hashing and found bugs that need to be solved.
the zobrist table is needed for the alpha beta search. the alpha beta search is called by the get_best_move method. When the get best move method in the client.py calls the alpha beta, it can't find the zobrist table as the variable does not persist to the other document. That needs to be solved